Moved files to new locations within the editor
