+++ Source/PokemonDataEditor/PokemonDataEditor.Build.cs
@@ -1,0 +1,30 @@
+﻿using UnrealBuildTool;
+
+public class PokemonDataEditor : ModuleRules
+{
+    public PokemonDataEditor(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
+
+        PublicDependencyModuleNames.AddRange(
+            new string[]
+            {
+                "Core",
+            }
+        );
+
+        PrivateDependencyModuleNames.AddRange(
+            new string[]
+            {
+                "CoreUObject",
+                "Engine",
+                "Slate",
+                "SlateCore",
+                "PokemonData",
+                "BlueprintGraph",
+                "KismetCompiler",
+                "UnrealEd"
+            }
+        );
+    }
+}
