+++ Source/PokemonDataEditor/Private/PokemonDataEditor.cpp
@@ -1,0 +1,15 @@
+﻿#include "PokemonDataEditor.h"
+
+#define LOCTEXT_NAMESPACE "FPokemonDataEditorModule"
+
+void FPokemonDataEditorModule::StartupModule() {
+    
+}
+
+void FPokemonDataEditorModule::ShutdownModule() {
+    
+}
+
+#undef LOCTEXT_NAMESPACE
+    
+IMPLEMENT_MODULE(FPokemonDataEditorModule, PokemonDataEditor)
