+++ Source/PokemonDataEditor/Public/PokemonDataEditor.h
@@ -1,0 +1,13 @@
+﻿#pragma once
+
+#include "CoreMinimal.h"
+#include "Modules/ModuleManager.h"
+
+/**
+ * Module for handling the data nodes in the editor
+ */
+class FPokemonDataEditorModule : public IModuleInterface {
+public:
+    virtual void StartupModule() override;
+    virtual void ShutdownModule() override;
+};
