+++ Source/UnrealPokemon.Target.cs
@@ -10,7 +10,6 @@ public class UnrealPokemonTarget : TargetRules
 		Type = TargetType.Game;
 		DefaultBuildSettings = BuildSettingsVersion.V4;
 		IncludeOrderVersion = EngineIncludeOrderVersion.Unreal5_3;
-		ExtraModuleNames.Add("UnrealPokemon");
 		RegisterModulesCreatedByRider();
 	}
 
@@ -19,7 +18,7 @@ public class UnrealPokemonTarget : TargetRules
 		ExtraModuleNames.AddRange(new string[]
 		{
 			"PokemonData", "PokemonEditorUtils", "PokemonUtilities", "PokemonCore", "GridBased2D", "GridBased2DEditor",
-			"RPGMenus", "PokemonUI", "RPGMenusEditor", "CommonMath", "PokemonTests"
+			"RPGMenus", "PokemonUI", "RPGMenusEditor", "CommonMath", "PokemonTests", "PokemonDataEditor"
 		});
 	}
 }
