--- Source/UnrealPokemon/UnrealPokemon.cpp
@@ -1,5 +1,0 @@
-// "Unreal Pokémon" created by Retro & Chill.
-#include "UnrealPokemon.h"
-#include "Modules/ModuleManager.h"
-
-IMPLEMENT_PRIMARY_GAME_MODULE(FDefaultGameModuleImpl, UnrealPokemon, "UnrealPokemon");
