--- Source/UnrealPokemon/UnrealPokemon.h
@@ -1,4 +1,0 @@
-// "Unreal Pokémon" created by Retro & Chill.
-#pragma once
-
-#include "CoreMinimal.h"
