+++ Source/UnrealPokemonEditor.Target.cs
@@ -9,7 +9,6 @@ public class UnrealPokemonEditorTarget : TargetRules
 		Type = TargetType.Editor;
 		DefaultBuildSettings = BuildSettingsVersion.V4;
 		IncludeOrderVersion = EngineIncludeOrderVersion.Unreal5_3;
-		ExtraModuleNames.Add("UnrealPokemon");
 		RegisterModulesCreatedByRider();
 	}
 
@@ -18,7 +17,7 @@ public class UnrealPokemonEditorTarget : TargetRules
 		ExtraModuleNames.AddRange(new string[]
 		{
 			"PokemonData", "PokemonEditorUtils", "PokemonUtilities", "PokemonCore", "GridBased2D", "GridBased2DEditor",
-			"RPGMenus", "PokemonUI", "RPGMenusEditor", "CommonMath", "PokemonTests"
+			"RPGMenus", "PokemonUI", "RPGMenusEditor", "CommonMath", "PokemonTests", "PokemonDataEditor"
 		});
 	}
 }
