+++ UnrealPokemon.uproject
@@ -5,11 +5,6 @@
 	"Description": "",
 	"Modules": [
 		{
-			"Name": "UnrealPokemon",
-			"Type": "Runtime",
-			"LoadingPhase": "Default"
-		},
-		{
 			"Name": "PokemonData",
 			"Type": "Runtime",
 			"LoadingPhase": "Default"
@@ -61,6 +56,11 @@
 		},
 		{
 			"Name": "PokemonTests",
+			"Type": "Editor",
+			"LoadingPhase": "Default"
+		},
+		{
+			"Name": "PokemonDataEditor",
 			"Type": "Editor",
 			"LoadingPhase": "Default"
 		}
